requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
docopt==0.6.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
iso8601==1.1.0
multidict==6.0.4
numpy==1.22.4
openai==0.27.7
Pillow==8.4.0
pipreqs==0.4.13
pyserial==3.5
PyYAML==6.0
requests==2.30.0
serial==0.0.97
streamlit==1.25.0
tensorflow==2.12.0
tqdm==4.65.0
urllib3==2.0.2
yarg==0.1.9
yarl==1.9.2
